Move FaceTime fallback into an extension to satisfy SwiftLint type_body_length
